requirements.txt
ipython>=7.29.0
jupyterlab>=3.2.3
matplotlib>=3.5.0
numpy>=1.19.5
pandas>=1.2.4
Pillow>=8.4.0
tensorflow>=2.11.0
tensorflow_datasets>=4.3.0
tensorflow_io>=0.21.0
torchjpeg>=0.9.21
torchvision>=0.10.0
tqdm>=4.61.1
psutil>=5.9.0
tensorflow_probability>=0.16.0
protobuf==3.20.3
advex-uar>=0.0.5.dev0
tensorboardX>=2.0
scikit-image
lpips
onnx
onnx_tf
wget
opencv-python
